Move the self-test pin to the class that owns that surface (#2578 review)
Two review nits, both right.

The pin belongs in HeadlessSelfTestTests, which documents itself as pinning
this flag's contracts. I had dropped it into ViewerStoreUnreachableTests, whose
class doc scopes it to distinguishing an unreachable store from a read-only
seat - unrelated to reading HeadlessSelfTest.cs's source.

And the usings are kept alphabetical here; System.Runtime.CompilerServices sorts
after System.IO, not after the project namespaces.
