+++ Darling/Darling.Tests/ViewerDataServiceTests.cs
@@ -9,7 +9,6 @@
 using System;
 using System.Globalization;
 using System.IO;
-using System.Runtime.CompilerServices;
 using System.Linq;
 using System.Text.Json;
 using System.Threading.Tasks;
@@ -735,45 +734,6 @@ public sealed class ViewerStoreUnreachableTests
            detail cost a field operator hours, and the fixed prose alone reads identically for a TLS chain
            rejection, a wrong password, a pg_hba refusal, and a dead host. */
         Assert.Contains("Underlying error: connection refused", ex.Message, StringComparison.Ordinal);
-    }
-
-    private static string RepoRoot([CallerFilePath] string thisFile = "")
-    {
-        var dir = Path.GetDirectoryName(thisFile)!;
-        while (dir is not null && !File.Exists(Path.Combine(dir, "PerformanceMonitor.sln")) && !Directory.Exists(Path.Combine(dir, ".git")))
-        {
-            dir = Path.GetDirectoryName(dir);
-        }
-
-        Assert.NotNull(dir);
-        return dir!;
-    }
-
-    /// <summary>
-    /// #2578: the self-test must probe the connection string STARTUP opens, not the raw darling.json one.
-    ///
-    /// <para>MainWindow builds its data service as
-    /// <c>new ViewerDataService(settings.ConnectionString, appSettings.ConnectionTimeoutSeconds)</c>, and that
-    /// constructor rewrites the string through <see cref="ViewerDataService.ApplyConnectionTimeout"/>. The
-    /// headless probe used the raw string, so the two diverged whenever a Connection timeout preference
-    /// existed — which is always, since <see cref="ViewerAppSettings.ConnectionTimeoutSeconds"/> defaults to 5
-    /// and is clamped to 5..60.</para>
-    ///
-    /// <para>The consequence is a diagnostic that can pass every layer while startup fails, which is the
-    /// report that found it. Asserted against the shipped source: the probe must not be handed
-    /// <c>settings.ConnectionString</c> directly.</para>
-    /// </summary>
-    [Fact]
-    public void TheHeadlessSelfTest_ProbesTheEffectiveConnectionString_NotTheRawOne()
-    {
-        var source = File.ReadAllText(Path.Combine(
-            RepoRoot(), "Darling", "PerformanceMonitor.Darling.Viewer", "HeadlessSelfTest.cs"));
-
-        Assert.DoesNotContain(
-            "StoreConnectionSelfTest.RunAsync(settings.ConnectionString)", source, StringComparison.Ordinal);
-        Assert.Contains(
-            "StoreConnectionSelfTest.RunAsync(effectiveConnectionString)", source, StringComparison.Ordinal);
-        Assert.Contains("ViewerDataService.ApplyConnectionTimeout(", source, StringComparison.Ordinal);
     }
 
     /// <summary>
